feat. 카페 리뷰 작성자에게만 삭제 버튼 보이도록 처리

+++ src/main/resources/static/js/cafe/cafeReview.js
@@ -51,9 +51,13 @@ function getCafeReviewList() {
         callback: function (data, pagination) {
             $('#review-list-container').empty();
             for (let review of data) {
-                // console.log(review)
                 let tempHtml = makeHtmlReview(review);
+                let userId = review["userId"];
+                let cafeReviewNumber = review["cafeReviewNumber"];
                 $('#review-list-container').append(tempHtml);
+                if (loginUserId !== userId){
+                    $(`.delete${cafeReviewNumber}`).hide();
+                }
             }
         }
     });
@@ -80,8 +84,9 @@ function makeHtmlReview(review) {
                             <div class="review-info-middle">${cafeReviewContent}</div>
                             <div class="review-info-bottom">
                                 <div class="review-create-date">${createdDate}</div>
-                                <div>
-                                    <button onclick="deleteReview(${cafeReviewNumber})" id="review-del-btn">삭제</button>
+                                <div class="form-group">
+                                    <script>alert("loginUserId = " + ${loginUserId} + "userId =" + ${userId})</script>
+                                    <button onclick="deleteReview(${cafeReviewNumber})" class="delete${cafeReviewNumber}" id="review-del-btn">삭제</button>
                                 </div>
                             </div>
                         </div>
@@ -89,12 +94,14 @@ function makeHtmlReview(review) {
 }
 
 function deleteReview(cafeReviewNumber) {
-    $.ajax({
-        type: "DELETE",
-        url: `/api/cafes/review/${cafeReviewNumber}`,
-        data: {},
-        success: function (response) {
-            getCafeReviewList();
-        }
-    })
+    if(confirm("삭제 하시겠습니까?")){
+        $.ajax({
+            type: "DELETE",
+            url: `/api/cafes/review/${cafeReviewNumber}`,
+            data: {},
+            success: function (response) {
+                getCafeReviewList();
+            }
+        })
+    }
 }
